+++ CHANGES.md
@@ -11,7 +11,8 @@
 * Add `ResponseError` impl for `SendRequestError`.
   This improves ergonomics of http client.
   
-* Allow connection timeout to be set
+* Allow client connection timeout to be set #108
+
 
 ## 0.4.4 (2018-03-04)